requirements.txt
aiohttp
python-dotenv
rich